major updates to AST construction, see verbose message
this update changes the %type for many rules from TreeNode to a more specific subclass of TreeNode. It changes the type of the 'profilelist' rule from 'TreeNode' to a pointer of 'std::list<ProfileNode>'. It improves how start/stop character position are calculated for the 'rules' rule, and creates the 'RulesList' class for that rule. Also we are now saving the PrefixNode for many rules.